api-ms-agent/pyproject.toml
[project]
name = "api-ms-agent"
version = "0.1.0"
description = "FastAPI backend with Microsoft Agent Framework for chat functionality"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agent-framework>=1.0.0b251120",
    "azure-ai-documentintelligence>=1.0.0",
    "azure-cosmos>=4.9.0",
    "azure-identity>=1.25.1",
    "fastapi>=0.122.0",
    "openai>=2.8.1",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "structlog>=25.5.0",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.7",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
